All threads appear to correctly communicate with each other using mutexes and condition variables. Test function appears to work but is not terminating.

+++ CSUBatch/Analytics.h
@@ -9,14 +9,13 @@
 #ifndef CSUBATCH_ANALYTICS_H
 #define CSUBATCH_ANALYTICS_H
 
-//#include <bits/types/time_t.h>
-#include <sys/types.h>
+#include <time.h>
 
 time_t starting_time;
 int total_number_of_jobs;
-float turnaround_time;
-float CPU_time;
-float waiting_time;
+long turnaround_time;
+long CPU_time;
+long waiting_time;
 float throughput;
 time_t job_start_time;
